+++ packages/showcases/atomic-components/.storybook/preview.js
@@ -2,7 +2,7 @@
 import { withDesign } from 'storybook-addon-designs';
 import backgrounds from '@acko-components/showcase-core/addons/backgrounds.json';
 import viewports from '@acko-components/showcase-core/addons/viewports.json';
-import '@acko-components/icons/dist/vitamix/font/vitamix.css';
+import '@acko-components/icons/dist/ackox/font/ackox.css';
 
 export const decorators = [withDesign];