+++ packages/showcases/css/.storybook/preview.js
@@ -1,5 +1,5 @@
 import '@acko-components/css';
-import '@acko-components/icons/dist/vitamix/font/vitamix.css';
+import '@acko-components/icons/dist/ackox/font/ackox.css';
 import { withDesign } from 'storybook-addon-designs';
 import backgrounds from '@acko-components/showcase-core/addons/backgrounds.json';
 import viewports from '@acko-components/showcase-core/addons/viewports.json';