+++ packages/sources/icons/src/scripts/add-sprite-declaration.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 
 fs.writeFileSync(
-  `${process.cwd()}/dist/vitamix/sprite/vitamix.svg.d.ts`,
+  `${process.cwd()}/dist/ackox/sprite/ackox.svg.d.ts`,
   'export default SVGElement;',
 );