+++ packages/sources/icons/CHANGELOG.md
@@ -113,4 +113,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Features
 
-* **@acko-components/icons:** add Vitamix Icons library (from Remix Icon) ([#189](https://github.com/Decathlon/vitamin-web/issues/189)) ([64acac5](https://github.com/Decathlon/vitamin-web/commit/64acac5c51b24275ce121c24d79f3bfdcecd748c))
+* **@acko-components/icons:** add ackox Icons library (from Remix Icon) ([#189](https://github.com/Decathlon/vitamin-web/issues/189)) ([64acac5](https://github.com/Decathlon/vitamin-web/commit/64acac5c51b24275ce121c24d79f3bfdcecd748c))